I implemented join game using websocket, but I was supposed to join in the usual way and then open a websocket connection.

+++ shared/src/main/requests/JoinGameRequest.java
@@ -27,7 +27,7 @@ public class JoinGameRequest extends Request {
     }
 
     public Integer getGameID() {
-        return (Integer) joinGameRequest.get("gameID");
+        return ((Double) joinGameRequest.get("gameID")).intValue();
     }
 
     public String getAuthToken() {
